[22/n] tensor engine, Move Endpoint to its own file, move propagator to endpoint
This splits Endpoint into its own file since it is gaining more functionality.
Propagator rules, which previously were attached to Remote, a subclass of Endpoint, are now defined on Endpoint. The @endpoint annotation can now take a propagator argument just like a remote. Since some propagators require the tensor engine, this introduces a shim through which that functionality can be accesssed.

Differential Revision: [D78771395](https://our.internmc.facebook.com/intern/diff/D78771395/)

[ghstack-poisoned]
