+++ python/monarch/_src/actor/code_sync/auto_reload.py
@@ -17,7 +17,8 @@ from pathlib import Path
 from types import ModuleType
 from typing import Dict, List, Optional, Tuple
 
-from monarch._src.actor.actor_mesh import Actor, endpoint
+from monarch._src.actor.actor_mesh import Actor
+from monarch._src.actor.endpoint import endpoint
 
 
 class SysAuditHookGuard(contextlib.AbstractContextManager):
