+++ python/monarch/actor/__init__.py
@@ -15,13 +15,13 @@ from monarch._src.actor.actor_mesh import (
     current_actor_name,
     current_rank,
     current_size,
-    endpoint,
     MonarchContext,
     Point,
     port,
     send,
     ValueMesh,
 )
+from monarch._src.actor.endpoint import endpoint
 from monarch._src.actor.future import Future
 from monarch._src.actor.proc_mesh import (
     debug_client,
